feat: refactor authentication to use UserLoginDTO and update user service methods
