+++ app-goconecta/app-goconecta.Server/DTOs/UserLoginDTO.cs
@@ -1,0 +1,7 @@
+namespace app_goconecta.Server.DTOs;
+
+public class UserLoginDTO
+{
+    public string Email { get; set; }
+    public string Password { get; set; }
+}
